requirements.txt
asgiref==3.3.1
boto3==1.16.36
botocore==1.19.36
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0
django-cors-headers==3.5.0
django-heroku==0.3.1
django-sendgrid-v5==0.9.0
django-storages==1.10.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
packaging==20.7
Pillow==8.0.1
psycopg2==2.7.5
psycopg2-binary==2.8.6
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-http-client==3.3.1
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
sendgrid==6.4.8
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0